small renames, moved mid360 realsense to robot/assembly
