--- dimos/hardware/sensors/lidar/mid360_realsense_30/recorder.py
@@ -1,39 +1,0 @@
-# Copyright 2026 Dimensional Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Records the RealSense D435i + Mid-360 rig into a memory2 SQLite db.
-
-Captures Point-LIO odom + lidar (trajectory baked into ``pointlio_lidar`` via the
-inherited ``@pose_setter_for``) plus the RealSense color/depth/pointcloud streams. The
-raw Livox stream is NOT recorded here — enable the pcap recorder in the record blueprint
-to capture it. Companion streams are recorded as-is and anchored via the static mount
-frames published on tf.
-"""
-
-from __future__ import annotations
-
-from dimos.core.stream import In
-from dimos.hardware.sensors.lidar.pointlio.pose_recorder import PointlioPoseRecorder
-from dimos.msgs.sensor_msgs.CameraInfo import CameraInfo
-from dimos.msgs.sensor_msgs.Image import Image
-from dimos.msgs.sensor_msgs.PointCloud2 import PointCloud2
-
-
-class Mid360RealsenseRecorder(PointlioPoseRecorder):
-    # pointlio_odometry / pointlio_lidar are inherited from PointlioPoseRecorder.
-    color_image: In[Image]
-    realsense_depth_image: In[Image]
-    realsense_pointcloud: In[PointCloud2]
-    realsense_camera_info: In[CameraInfo]
-    realsense_depth_camera_info: In[CameraInfo]
