+++ dimos/robot/all_blueprints.py
@@ -73,8 +73,8 @@ all_blueprints = {
     "mid360-fastlio-voxels-native": "dimos.hardware.sensors.lidar.fastlio2.fastlio_blueprints:mid360_fastlio_voxels_native",
     "mid360-pointlio": "dimos.hardware.sensors.lidar.pointlio.pointlio_blueprints:mid360_pointlio",
     "mid360-pointlio-voxels": "dimos.hardware.sensors.lidar.pointlio.pointlio_blueprints:mid360_pointlio_voxels",
-    "mid360-realsense-record": "dimos.hardware.sensors.lidar.mid360_realsense_30.blueprints:mid360_realsense_record",
-    "mid360-realsense-record-with-pcap": "dimos.hardware.sensors.lidar.mid360_realsense_30.blueprints:mid360_realsense_record_with_pcap",
+    "mid360-realsense-record": "dimos.robot.assembly.mid360_realsense_30:mid360_realsense_record",
+    "mid360-realsense-record-with-pcap": "dimos.robot.assembly.mid360_realsense_30:mid360_realsense_record_with_pcap",
     "openarm-mock-planner-coordinator": "dimos.robot.manipulators.openarm.blueprints.planner:openarm_mock_planner_coordinator",
     "openarm-planner-coordinator": "dimos.robot.manipulators.openarm.blueprints.planner:openarm_planner_coordinator",
     "path-planner-eval": "dimos.navigation.nav_3d.evaluator.blueprints:path_planner_eval",
@@ -196,8 +196,8 @@ all_modules = {
     "mcp-server": "dimos.agents.mcp.mcp_server.McpServer",
     "memory-module": "dimos.memory2.module.MemoryModule",
     "mid360-pcap-recorder": "dimos.hardware.sensors.lidar.virtual_mid360.recorder.Mid360PcapRecorder",
-    "mid360-realsense-recorder": "dimos.hardware.sensors.lidar.mid360_realsense_30.recorder.Mid360RealsenseRecorder",
-    "mid360-realsense-static-tf": "dimos.hardware.sensors.lidar.mid360_realsense_30.static_transforms.Mid360RealsenseStaticTf",
+    "mid360-realsense-recorder": "dimos.robot.assembly.mid360_realsense_30.Mid360RealsenseRecorder",
+    "mid360-realsense-static-tf": "dimos.robot.assembly.mid360_realsense_30.Mid360RealsenseStaticTf",
     "mls-planner-native": "dimos.navigation.nav_3d.mls_planner.mls_planner_native.MLSPlannerNative",
     "mock-b1-connection-module": "dimos.robot.unitree.b1.connection.MockB1ConnectionModule",
     "module-a": "dimos.robot.unitree.demo_error_on_name_conflicts.ModuleA",
